优化项目依赖和代码结构
1. 清理go.mod和go.sum，移除不必要的gorm和sqlite依赖
2. 调整strval.go文件注释格式，提高可读性
3. 更新strval_db_test.go测试文件
4. 添加新的strval_string_test.go测试文件
5. 更新README.md文档

所有测试已通过，确保库的功能完整性和稳定性

+++ strval_db_test.go
@@ -84,7 +84,7 @@ func TestIntDatabase(t *testing.T) {
 	if err != nil {
 		t.Errorf("Int.Value() returned error: %v", err)
 	}
-	if intVal, ok := val.(int); !ok || intVal != 42 {
+	if intVal, ok := val.(int64); !ok || intVal != 42 {
 		t.Errorf("Int.Value() returned %v, want 42", val)
 	}
